refactor(trusty-agents-common): extract TokenUsage/PerfRecord value types (refs #830)
Wave 2 Part A of the trusty-agents-common build-out (issue #867).

## What moved

Four portable plain-data types from `crates/trusty-agents/src/perf/` into
a new `crates/trusty-agents-common/src/perf.rs` module:

- `TokenUsage` — 4-field token counter (prompt/completion/cache_read/cache_creation)
- `PhaseRecord` — per-phase duration + token snapshot
- `PerfTotals` — rolled-up totals across all phases
- `PerfRecord` — full run record (persisted to docs/performance/runs/)

`PerfCollector` (stateful: file I/O, timing, Tokio async) stays in
`trusty-agents::perf` — only the portable value types move.

## Compat shim

`trusty-agents/src/perf/mod.rs` replaces the four struct definitions with
EXPLICIT re-exports (no wildcards):

    pub use trusty_agents_common::perf::{PerfRecord, PerfTotals, PhaseRecord, TokenUsage};

Every existing `crate::perf::TokenUsage` / `PhaseRecord` / `PerfTotals` /
`PerfRecord` reference inside trusty-agents continues to resolve unchanged.

## New deps in trusty-agents-common

- `serde` already present; no new deps needed for perf.rs

## Verification

- No `crate::` back-references in `trusty-agents-common/src/perf.rs`
- `cargo build --workspace` → clean
- `cargo test --workspace` → 7123 passed, 0 failed
- `cargo clippy --workspace --all-targets -- -D warnings` → clean
- `cargo fmt --check` → clean
- `bash scripts/check_line_cap.sh` → 0 violations
